Refactoring QrCode verification component to make it reusable
