+++ components/qr-generator.jsx
