+++ components/qr-reader.jsx
